Add reset_status button

+++ view/ticket_database_view.py
@@ -61,16 +61,19 @@ class DatabaseMenuButtons(discord.ui.View):
 
         cancel_reservetion_button = discord.ui.Button(label="Отменить бронь", emoji="🚩", style=discord.ButtonStyle.gray, custom_id="database_menu_cancel_reservetion", row=2)
         add_reservetion_button = discord.ui.Button(label="Добавить бронь", emoji="🚩", style=discord.ButtonStyle.gray, custom_id="database_menu_add_reservetion", row=2)
+        reset_status = discord.ui.Button(label="Обновить статус места", emoji="🔄", style=discord.ButtonStyle.gray, custom_id="database_menu_reset_status", row=2)
 
         back_button.callback = self.back_callback
         next_button.callback = self.next_callback
         cancel_reservetion_button.callback = self.cancel_reservetion_callback
         add_reservetion_button.callback = self.add_reservetion_callback
+        reset_status.callback = self.reset_status_callback
 
         self.add_item(back_button)
         self.add_item(next_button)
         self.add_item(cancel_reservetion_button)
         self.add_item(add_reservetion_button)
+        self.add_item(reset_status)
 
     async def back_callback(self, interaction: discord.Interaction):
         embeds = DatabaseMenuEmbed(self.floor - 1).get_seat_list()
@@ -81,39 +84,48 @@ class DatabaseMenuButtons(discord.ui.View):
         await interaction.response.edit_message(embeds=embeds, view=DatabaseMenuButtons(self.floor+1))
 
     async def cancel_reservetion_callback(self, interactrion: discord.Interaction):
-        await interactrion.response.send_message(view=SeatSelectView(interactrion.message, floor=self.floor), ephemeral=True)
+        await interactrion.response.send_message(view=SeatSelectView(interactrion.message, floor=self.floor, remove_seat=True), ephemeral=True)
 
     async def add_reservetion_callback(self, interactrion: discord.Interaction):
         await interactrion.response.send_message(view=SeatSelectView(interactrion.message, floor=self.floor, add_seat=True), ephemeral=True)
 
+    async def reset_status_callback(self, interactrion: discord.Interaction):
+        await interactrion.response.send_message(view=SeatSelectView(interactrion.message, floor=self.floor, reset_status=True), ephemeral=True)
+
 
 class DatabaseMenuSeatSelect(discord.ui.Select):
-    def __init__(self, message, floor: int, add_seat: bool = False):
+    def __init__(self, message, floor: int, remove_seat: bool = False,  add_seat: bool = False, reset_status: bool = False):
         options = []
 
         self.db = TicketBookingDatabase()
 
         self.floor = floor
         self.message = message
+        self.remove_seat = remove_seat
         self.add_seat = add_seat
+        self.reset_status = reset_status
 
         for seat in self.db.get_seat_list(floor):
-            emoji = ""
-            if seat[1] is None:
-                emoji = "⚫"
-            else:
-                emoji = "🔵"
+            status = "✅"
+            avalible = self.db.is_avalible(self.floor, seat[0])
+            if not avalible:
+                if seat[1] is None:
+                    status = "🟨"
+                else:
+                    status = "🟥"
 
             options.append(
                 discord.SelectOption(
                     label=seat[0],
                     value=f'{seat[0]}_{seat[1]}',
                     description=f"{seat[1]} | {seat[2]}",
-                    emoji=emoji
+                    emoji=status
                 )
             )
-
-        placeholder = {False: "Удалить бронь...", True: "Добавить бронь..."}
+        if not reset_status:
+            placeholder = {False: "Удалить бронь...", True: "Добавить бронь..."}
+        else:
+            placeholder = {False: "Обновить статус..."}
 
         super().__init__(
             custom_id='seat_select',
@@ -122,23 +134,33 @@ class DatabaseMenuSeatSelect(discord.ui.Select):
         )
 
     async def callback(self, inter: discord.Interaction):
-        if self.values[0].split("_")[1] != "None":
-            seat = self.values[0].split("_")[0]
+        seat = self.values[0].split("_")[0]
+
+        if self.remove_seat and self.values[0].split("_")[1] != "None":
             self.db.remove_user(self.floor, seat)
             self.db.set_seat_status(self.floor, seat, SeatStatus.AVAILABLE)
 
             embed = DatabaseMenuEmbed(self.floor).get_seat_list()
             await self.message.edit(embeds=embed, view=DatabaseMenuButtons(floor=self.floor))
             await inter.response.edit_message(view=SeatSelectView(self.message, self.floor))
-        elif self.add_seat:
-            await inter.response.edit_message(view=UserSelectView(self.message, self.floor, self.values[0].split("_")[0]))
+
+        elif self.add_seat and self.values[0].split("_")[1] == "None":
+            await inter.response.edit_message(view=UserSelectView(self.message, self.floor, seat))
+
+        elif self.reset_status and self.values[0].split("_")[1] == "None" and not self.db.is_avalible(self.floor, seat):
+            self.db.set_seat_status(self.floor, seat, SeatStatus.AVAILABLE)
+            embed = DatabaseMenuEmbed(self.floor).get_seat_list()
+            await self.message.edit(embeds=embed, view=DatabaseMenuButtons(floor=self.floor))
+            await inter.response.edit_message(view=SeatSelectView(self.message, self.floor, reset_status=True))
+        else:
+            await inter.response.send_message(embed=discord.Embed(title="О нет, ашипка!", description="Кажется вы что-то делаете не так(", color=discord.Color.red()), ephemeral=True, delete_after=3)
 
 
 class SeatSelectView(discord.ui.View):
-    def __init__(self, message, floor: int, add_seat: bool = False):
+    def __init__(self, message, floor: int, remove_seat: bool = False, add_seat: bool = False, reset_status: bool = False):
         super().__init__(timeout=None)
 
-        self.add_item(DatabaseMenuSeatSelect(message=message, floor=floor, add_seat=add_seat))
+        self.add_item(DatabaseMenuSeatSelect(message=message, floor=floor, remove_seat=remove_seat, add_seat=add_seat, reset_status=reset_status))
 
 
 class UserSelect(discord.ui.UserSelect):
